requirements.txt
pbr>=0.5.21,<1.0
Babel>=0.9.6
python-swiftclient>=1.6
python-keystoneclient>=0.4.2
python-magic
PyYAML
